refactor(services): split base/world_entry.rs into submodules
Splits the 1,067-line world_entry.rs orchestrator into focused siblings
(play_character, enable_entities, map_loaded, gate_travel, cell_dispatch,
space_registry) under base/world_entry/.

Consolidates world_entry_methods/ under world_entry/methods/.
Public API unchanged. No behavior changes.
