+++ crates/services/src/base/world_entry/methods/inventory/core.rs
