+++ crates/services/src/base/world_entry/methods/inventory/move_.rs
