+++ crates/services/src/base/world_entry/methods/mail.rs
