+++ crates/services/src/base/world_entry/methods/progression.rs
