+++ crates/services/src/base/world_entry/play_character.rs
@@ -1,0 +1,141 @@
+//! `playCharacter` -> entity teardown step.
+//!
+//! When the client picks a character and calls `playCharacter`, we tear down
+//! its previous entity tree by sending RESET_ENTITIES. The next packet from
+//! the client is ENABLE_ENTITIES, which drives the create-player step in
+//! `enable_entities.rs`.
+
+use std::collections::HashMap;
+use std::net::SocketAddr;
+use std::sync::atomic::Ordering;
+use std::sync::{Arc, Mutex};
+
+use sqlx::PgPool;
+use tokio::net::UdpSocket;
+use tokio::sync::mpsc;
+
+use cimmeria_entity::manager::EntityManager;
+
+use crate::cell::messages::BaseToCellMsg;
+use crate::mercury::build_reset_entities;
+
+use super::super::ConnectedClientState;
+use super::methods::{query_player_load_data, query_world_entry};
+
+/// Send RESET_ENTITIES when the client calls `playCharacter` to begin world entry.
+pub(crate) async fn handle_play_character(
+    socket: &Arc<UdpSocket>,
+    addr: SocketAddr,
+    key: [u8; 32],
+    account_id: u32,
+    player_id: i32,
+    connected: &Arc<Mutex<HashMap<SocketAddr, ConnectedClientState>>>,
+    db_pool: &Option<Arc<PgPool>>,
+    entity_manager: &Arc<Mutex<EntityManager>>,
+    cell_tx: &Option<mpsc::Sender<BaseToCellMsg>>,
+) -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
+    // Guard: only send once per connection.
+    let arcs = {
+        let mut clients = connected
+            .lock()
+            .map_err(|_| "connected lock poisoned")?;
+        if let Some(c) = clients.get_mut(&addr) {
+            if !c.world_entry_sent {
+                c.world_entry_sent = true;
+                Some((Arc::clone(&c.pending_acks), Arc::clone(&c.next_seq)))
+            } else {
+                None
+            }
+        } else {
+            tracing::warn!(%addr, "play_character: addr not in connected map");
+            None
+        }
+    };
+
+    let (pending_acks_arc, next_seq) = match arcs {
+        Some(a) => a,
+        None => return Ok(()),
+    };
+
+    // Query character data from DB and resolve space via CellService
+    let entry_info = query_world_entry(db_pool, account_id, player_id, entity_manager, cell_tx).await;
+
+    // query_world_entry returns NO_ENTITY_ID as a "world entry failed"
+    // sentinel (DB error or character not found). Bail before dispatching any
+    // packets that would target an entity the cell never registered, AND
+    // clear the world_entry_sent latch so a subsequent playCharacter retry on
+    // the same connection isn't silently dropped.
+    if entry_info.player_entity_id == super::methods::world_entry_db::NO_ENTITY_ID {
+        tracing::error!(
+            %addr, player_id, account_id,
+            "World entry aborted: query_world_entry returned NO_ENTITY_ID sentinel"
+        );
+        if let Ok(mut clients) = connected.lock() {
+            if let Some(c) = clients.get_mut(&addr) {
+                c.world_entry_sent = false;
+            }
+        }
+        return Ok(());
+    }
+
+    // Also query the full player data needed for mapLoaded
+    let player_load_data = query_player_load_data(db_pool, account_id, player_id).await;
+
+    // NOTE: C++ Account.py:293-296 uses SGWGmPlayer (0x03) for access_level > 0,
+    // but SGWGmPlayer adds 6 ClientMethods and 80+ CellMethods that shift ALL
+    // flattened method indices. Our hardcoded method_idx constants (BeingAppearance=26,
+    // etc.) only work for SGWPlayer. Until we build a separate SGWGmPlayer index
+    // table, always use SGWPlayer (0x02) regardless of access_level.
+    // TODO: Build SGWGmPlayer method index table to enable GM entity type.
+
+    tracing::info!(
+        %addr,
+        player_id,
+        entity_id = entry_info.player_entity_id,
+        space_id = entry_info.space_id,
+        pos = ?entry_info.pos,
+        class_id = entry_info.class_id,
+        "World entry: sending RESET_ENTITIES (entity teardown)"
+    );
+
+    let acks: Vec<u32> = {
+        let mut pending = pending_acks_arc.lock().unwrap();
+        pending.drain(..).collect()
+    };
+
+    // Entity teardown: Send ONLY RESET_ENTITIES.
+    // The C++ server sends RESET_ENTITIES in its own flushed bundle. The client
+    // tears down all entities, then sends ENABLE_ENTITIES, which triggers
+    // the create-player step (CREATE_BASE_PLAYER + viewport + cell + forced position).
+    let seq = next_seq.fetch_add(1, Ordering::Relaxed);
+    let pkt = build_reset_entities(&key, seq, &acks);
+    tracing::trace!(%addr, len = pkt.len(), seq, "UDP_OUT RESET_ENTITIES (entity teardown)");
+    socket.send_to(&pkt, addr).await?;
+
+    // Store the world entry info and player load data for the create-player step.
+    {
+        let mut clients = connected.lock().map_err(|_| "connected lock poisoned")?;
+        if let Some(c) = clients.get_mut(&addr) {
+            c.pending_player_entity_id = Some(entry_info.player_entity_id);
+            c.player_entity_id = Some(entry_info.player_entity_id);
+            c.player_name = Some(player_load_data.player_name.clone());
+            c.player_level = Some(player_load_data.level);
+            c.player_archetype = Some(player_load_data.archetype);
+            c.world_name = Some(entry_info.world_name.clone());
+            c.player_xp = Some(player_load_data.exp as u64);
+            c.player_training_points = Some(player_load_data.training_points as u32);
+            // Cache the selected player's DB id so gate-travel/respawn can use
+            // it directly instead of falling back to a "lowest player_id for
+            // this account" lookup that would target the wrong character on
+            // multi-character accounts.
+            c.active_player_id = Some(player_id);
+            c.pending_world_entry = Some(entry_info);
+            c.pending_player_load_data = Some(player_load_data);
+            c.pending_client_ready = None;
+        }
+    }
+
+    tracing::info!(%addr, "Entity teardown sent -- waiting for ENABLE_ENTITIES from client");
+
+    Ok(())
+}
